requirements.txt
streamlit
pandas
base58
matplotlib
numpy
yfinance
lxml
plotly
pillow
cufflinks